Use templates for postReconfigureScript and omnibusRubyConfig values
Switching to templates that just render out the value gives us
flexibility over how these values are rendered in parent charts.

For instance, in the gitlab-bundle chart, we will be able to
override the postReconfigureScript template to configure matter,

AND still add in any settings from the user's values file.

Same for the omnibusRubyConfig, we will be able to render out our
defaults, and still be able to tack on user's custom config.